requirements.txt
python-dotenv==1.1.1
openai==0.28